streamify everything
